+++ packages/node/src/indexer/blockDispatcher/stellar-block-dispatcher.ts
@@ -2,7 +2,9 @@
 // SPDX-License-Identifier: GPL-3.0
 
 import { IBlockDispatcher } from '@subql/node-core';
+import { StellarBlock } from '@subql/types-stellar';
 
-export interface IStellarBlockDispatcher extends IBlockDispatcher {
+export interface IStellarBlockDispatcher
+  extends IBlockDispatcher<StellarBlock> {
   init(onDynamicDsCreated: (height: number) => Promise<void>): Promise<void>;
 }